+++ src/main/java/org/apiphany/security/tls/RecordHeader.java
@@ -6,7 +6,6 @@ import java.nio.ByteBuffer;
 
 import org.apiphany.io.UInt16;
 import org.apiphany.io.UInt8;
-import org.apiphany.json.JsonBuilder;
 import org.apiphany.security.ssl.SSLProtocol;
 
 /**
@@ -22,7 +21,7 @@ public class RecordHeader implements TLSObject {
 	/**
 	 * The fixed size of a record header in bytes.
 	 */
-	public static final int BYTES = 5;
+	public static final int BYTES = RecordContentType.BYTES + Version.BYTES + UInt16.BYTES;
 
 	/**
 	 * The type of content in the record.
@@ -109,7 +108,7 @@ public class RecordHeader implements TLSObject {
 	 */
 	@Override
 	public String toString() {
-		return JsonBuilder.toJson(this);
+		return TLSObject.serialize(this);
 	}
 
 	/**
@@ -119,7 +118,7 @@ public class RecordHeader implements TLSObject {
 	 */
 	@Override
 	public int sizeOf() {
-		return type.sizeOf() + version.sizeOf() + length.sizeOf();
+		return BYTES;
 	}
 
 	/**